telegram-join-approval-bot/requirements.txt
python-telegram-bot[webhooks]>=22,<23
rapidocr-onnxruntime>=1.4,<2
opencc-python-reimplemented>=0.1.7,<1
Pillow>=10,<12
numpy>=1.26,<3
